Need to debug before it work

+++ public/adminTemplates/articles.html.php
@@ -2,9 +2,9 @@
 echo '<table>';
 foreach ($articles as $article) {
     echo '<tr>';
-    echo '<td>' . $article['title'] . '</td>';
+    echo '<td><a href="articledetail.php?id=' . $article['id'] . '">' . $article['title'] . '</a></td>';
     echo '<td><a href="editarticle.php?id=' . $article['id'] . '">Edit</a></td>';
-    echo '<td><a href="deletearticle.php?id=' . $article['id'] . '">Delete</a></td>';
+    echo '<td><a href="deletearticle.php?id=' . $article['id'] . '" onclick="return confirm(\'Are you sure you want to delete this user?\');">Delete</a></td>';
     echo '</td>';
 }
 echo '</table>';
