+++ public/adminTemplates/viewusers.html.php
@@ -6,7 +6,7 @@ foreach ($users as $user) {
     echo '<td>' . $user['name'] . '</td>';
     echo '<td>' . $user['status'] . '</td>';
     echo '<td><a href="register.php?id=' . $user['id'] . '">Edit</a></td>';
-    echo '<td><a href="deleteuser.php?id=' . $user['id'] . '">Delete</a></td>';
+    echo '<td><a href="deleteuser.php?id=' . $user['id'] . '" onclick="return confirm(\'Are you sure you want to delete this user?\');">Delete</a></td>';
     echo '</td>';
 }
 echo '</table>';
